+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Changelog
 * Update :command:`plot-bam-copy-number` command to run faster when ``--samples`` argument is used.
 * Change 'Unassigned' genotype to 'Indeterminate' genotype.
 * Add new method :meth:`api.core.get_variant_synonyms`.
+* Update :meth:`api.core.list_variants` method to accept multiple star alleles.
 
 0.5.0 (2021-10-02)
 ------------------